requirements.txt
Flask==2.3.2
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
flask-cors==4.0.0